# runner/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Livepeer AI Runner",
        "description": "An application to run AI pipelines",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "https://dream-gateway.livepeer.cloud",
            "description": "Livepeer Cloud Community Gateway"
        }
    ],
    "paths": {
...
        "/text-to-audio": {
            "post": {
                "summary": "Texttoaudio",
                "operationId": "TextToAudio",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Body_TextToAudio_text_to_audio_post"
                                    }
                                ],
                                "title": "Body"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "audio/mp4": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "detail": {
                                            "type": "string",
                                            "title": "Detail"
                                        }
                                    },
                                    "type": "object",
                                    "required": [
                                        "detail"
                                    ],
                                    "title": "HTTPError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "detail": {
                                            "type": "string",
                                            "title": "Detail"
                                        }
                                    },
                                    "type": "object",
                                    "required": [
                                        "detail"
                                    ],
                                    "title": "HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Body_TextToAudio_text_to_audio_post": {
                "properties": {
                    "text_input": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Text Input"
                    }
                },
                "type": "object",
                "title": "Body_TextToAudio_text_to_audio_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}